evals metadata in builtin JUnit reporter
These changes annotate the `testTask` in a way that the built-in JUnit reporter can
expose the evals metadata - similar to what is done for the JSON reporter.

The formatting is discussed at length [here](https://www.notion.so/sentry/Evals-Schema-for-JUnit-XML-2098b10e4b5d80609d62f5beffc3de26?source=copy_link)
JUnit is a very pervasive format, well established in the industry.